% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{grad.mix_2pl}
\alias{grad.mix_2pl}
\title{grad.mix_2pl Function}
\usage{
grad.mix_2pl(X, psi, z, nitems)
}
\arguments{
\item{X}{design matrix indexing each observation to corresponding item}

\item{psi}{vector of item parameters}

\item{z}{design matrix indexing each observation to corresponding individual}

\item{nitems}{integer especifying the number of items. It is used to separate parameters as psi[1:nitems] and psi[(nitems+1):(2*nitems)]}
}
\value{
a double vector
}
\description{
derivative of nlf_2pl with respect to theta
}
\examples{
data("test_data")
temp<-uirt_DM(test_data,"id")
nitems<-ncol(temp$X)
nind<-ncol(temp$Z)
alpha<-runif(nitems,0.6,1.5) ##simulation of discrimination parameters
beta<-runif(nitems,-2.4,2.4) ##simulation of dificulty parameters
theta<-rnorm(nind) ##simulation of random effects
psi<-c(-alpha*beta,log(alpha)) ###vector of parameters based on alpha and beta
res<-grad.mix_2pl(temp$X,psi,temp$Z,nitems) ###derivative of nlf_2pl with respect to theta evaluated at given values
}
\keyword{derivative}
