% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{inv_link_2pl}
\alias{inv_link_2pl}
\title{inv_link_2pl Function}
\usage{
inv_link_2pl(x)
}
\arguments{
\item{x}{a double vector}
}
\value{
a double vector
}
\description{
Inverse logit function, used as link in the 2pl model
}
\examples{
inv_link_2pl(rnorm(10))
}
\keyword{logit}
