% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{log_lik_2pl}
\alias{log_lik_2pl}
\title{log_lik_2pl Function}
\usage{
log_lik_2pl(y, X, psi, z, theta, nitems)
}
\arguments{
\item{y}{vectorised test}

\item{X}{design matrix indexing each observation to corresponding item}

\item{psi}{vector of item parameters}

\item{z}{design matrix indexing each observation to corresponding individual}

\item{theta}{vector of latent traits (i.e. random effects associated to individual level)}

\item{nitems}{integer especifying the number of items. It is used to separate parameters as psi[1:nitems] and psi[(nitems+1):(2*nitems)]}
}
\value{
a double atomic vector
}
\description{
log likelihood function for 2pl
}
\examples{
data("test_data")
temp<-uirt_DM(test_data,"id")
nitems<-ncol(temp$X)
nind<-ncol(temp$Z)
alpha<-runif(nitems,0.6,1.5) ##simulation of discrimination parameters
beta<-runif(nitems,-2.4,2.4) ##simulation of dificulty parameters
theta<-rnorm(nind) ##simulation of random effects
psi<-c(-alpha*beta,log(alpha)) ###vector of parameters based on alpha and beta
res<-log_lik_2pl(temp$Y,temp$X,psi,temp$Z,theta,nitems) ###2pl log likelihood evaluated at given values
}
\keyword{likelihood}
\keyword{log}
