% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{mh_gibbs_2pl}
\alias{mh_gibbs_2pl}
\title{mh_gibbs_2pl Function}
\usage{
mh_gibbs_2pl(dat, idname, psi0, theta0, B = diag(c(rep(9, dim(X)[2]), rep(1,
  dim(X)[2]))), b = c(rep(0, length(psi0))), Nc = 10000)
}
\arguments{
\item{dat}{a data frame containing test info. rows for individuals and columns for items. aditionally, a variable which identify individuals must be supied.}

\item{idname}{atomic character giving the name of the id variable in dat.}

\item{psi0}{a vector of initial values for parameter psi0. See nlf_2pl documentation.}

\item{theta0}{vector of initial values for latent traits. See nlf_2pl documentation.}

\item{B}{Prior covariance matrix for psi.}

\item{b}{Prior expected values for psi.}

\item{Nc}{number of mcmc iterations.}
}
\value{
a list with the following objects:
\itemize{
  \item fixp: a matrix where each row corresponds to an mcmc simulation of psi parameters.
  \item theta: a matrix where each row corresponds to an mcmc simulation of latent traits.
  \item Dev: a vector where deviance is calculated for each mcmc iteration.
  \item B: Prior covariance matrix used for psi.
  \item b: Prior Expected values used for psi.
}
}
\description{
this functions implements the metropolis-hastings within gibbs algorithm proposed in molano thesis.
}
\examples{
nitems<-7
nind<-100
sort(alpha_o<-runif(nitems,0.6,1.5))
sort(beta_o<-runif(nitems,-2.4,2.4))
d_o<- -alpha_o*beta_o

summary(theta_o<-rnorm(nind))
temp.b<-matrix(beta_o,ncol=nitems,nrow=nind, byrow =T)
temp.a<-matrix(alpha_o,ncol=nitems,nrow=nind, byrow =T)
temp.t<-matrix(theta_o,ncol=nitems,nrow=nind)
eta<-temp.a*(theta_o-temp.b)
pmod<-exp(eta)/(1+exp(eta))
test<-data.frame(ifelse(runif(nitems*nind)<pmod,1,0))
colnames(test)<-paste("t",1:nitems,sep="")
sort(apply(test,2,function(i)sum(i)*100/length(i)))
test<-data.frame(test,id=1:nind)
bres<-mh_gibbs_2pl(test,"id",psi0=c(d_o,log(alpha_o)),theta_o,Nc=1000)
}
\keyword{a}
\keyword{double}
\keyword{vector}
