% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{nlf_2pl}
\alias{nlf_2pl}
\title{nlf_2pl Function}
\usage{
nlf_2pl(X, psi, z, theta, nitems)
}
\arguments{
\item{X}{design matrix indexing each observation to corresponding item}

\item{psi}{vector of item parameters}

\item{z}{design matrix indexing each observation to corresponding individual}

\item{theta}{vector of latent traits (i.e. random effects associated to individual level)}

\item{nitems}{integer especifying the number of items. It is used to separate parameters as psi[1:nitems] and psi[(nitems+1):(2*nitems)]}
}
\value{
add style to sheet in wb
}
\description{
this functions is the non-linear function associated with the 2pl uirt model as estated in molano thesis.
}
\examples{
##none for now
}
\keyword{a}
\keyword{reals}
\keyword{vector}
\keyword{with}
