% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test_data data simulated for bayesuirt package tests}
\format{A data.frame object}
\usage{
data(test_data)
}
\description{
data simulated for bayesuirt tests
}
\examples{
data(test_data)
summary(test_data)
}
\keyword{datasets}
