% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesuirt.R
\name{uirt_DM}
\alias{uirt_DM}
\title{uirt_DM Function}
\usage{
uirt_DM(dat, idname)
}
\arguments{
\item{dat}{a data frame containing test info. rows for individuals and columns for items. aditionally, a variable which identify individuals must be supied.}

\item{idname}{atomic character giving the name of the id variable}
}
\value{
a list with the following objects:
\itemize{
  \item long.test: a data frame with id, item and y (0 or 1). Long format for the test.
  \item Y: Vector containing individual answers (0 or 1).
  \item X: desing matrix indexing each value of Y to the corresponding item.
  \item Z: desing matrix indexing each value of Y to the corresponding individual.
}
}
\description{
this functions converts a test dataframe to a vectoriced form following molano's thesis.
}
\examples{
data("test_data")
temp<-uirt_DM(test_data,"id")
}
\keyword{a}
\keyword{reals}
\keyword{vector}
\keyword{with}
